We can now display the toatal and quantity in the cart in the header
